api/requirements.txt
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Pillow==10.0.1
numpy==1.24.3